third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-invalid-names-expected.txt

ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character '@' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character 'b' in suborigin policy. Suborigin policy options must start and end with a single quote.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character 'F' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character 'F' in suborigin.
ALERT: I am a secret
